requirements.txt.txt
kivy==2.2.1
kivymd
pyrebase4
